Move gitlab-cli.ts into src directory
